backend/backend/requirements.txt
ansi2html==1.9.1
asgiref==3.7.2
blinker==1.7.0
certifi==2022.6.15
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.29.0
colorama==0.4.4
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dj-database-url==1.0.0
Django==4.2.8
django-appconf==1.0.6
django-autoslug==1.9.8
django-cities==0.6
django-cities-light==3.9.0
django-cookie-consent==0.4.0
django-cors-headers==4.3.1
django-countries==7.2.1
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-messages==0.6.0
django-model-utils==4.2.0
django-postman==4.2
django-smart-selects==1.5.9
djangorestframework==3.15.1
et-xmlfile==1.1.0
Flask==3.0.0
gunicorn==20.1.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.23.1
openpyxl==3.1.2
packaging==23.2
pandas==1.4.2
Pillow==9.0.0
plotly==5.18.0
progressbar2==4.0.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
python-utils==3.1.0
pytz==2021.3
requests==2.31.0
retrying==1.3.4
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tenacity==8.2.3
tqdm==4.62.3
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.2
urllib3==1.26.11
Werkzeug==3.0.1
whitenoise==6.2.0
zipp==3.17.0
